Fix: SSL.ZeroReturnError must raise ConnectionError
SSL.ZeroReturnError is raised when the connection has been closed cleanly,
and no more data will be received. This behavior indicates that the connection
is no longer usable and thus can be considered a type of connection error.
Therefore, it is appropriate to raise a ConnectionError when encountering SSL.ZeroReturnError.